Moved folders into apropiate layouts
