+++ app/containers/Cabinet/TablePage/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import TableDemo from '../../components/TableDemo';
+import TableDemo from '../../../components/TableDemo';
 
 export default class TablePage extends React.PureComponent {   // eslint-disable-line react/prefer-stateless-function
   render() {
