+++ app/containers/Default/LoginPage/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import NormalLoginForm from '../../components/LoginForm';
+import NormalLoginForm from '../../../components/LoginForm';
 
 export default class LoginPage extends React.PureComponent { // eslint-disable-line react/prefer-stateless-function
   render() {
